refactor(game): suffix input.cpp with _SDL2
